Fix cluster tracker: 64-bit timestamps, batch filter, polarity bug
- Change polarity::timestamp from int to int64_t to prevent overflow
  from caerPolarityEventGetTimestamp64 (root cause of clusters=0)
- Use findClusterNear() with fallback in batch filter instead of
  fastClusterFinder-only lookup, register new clusters in grid
- Fix addEvent() polarity logic: accept both polarities when
  useOnePolarityOnly is disabled (removed erroneous else branch)
- Remove forced useVelocity/thresholdVelocity override on enable
- Add cluster count debug overlay and periodic console logging

+++ src/ofxDVS.hpp
@@ -65,7 +65,7 @@
 struct polarity {
     int info;
     ofPoint pos;
-    int timestamp;
+    int64_t timestamp;
     bool pol;
     bool valid;
 };
